Flask_Project-2/requirements.txt
blinker==1.6.3
click==8.1.7
colorama==0.4.6
Flask==3.0.0
google-cloud-storage==2.12.0
google-cloud-datastore==2.18.0
google-cloud-secret-manager==2.16.3 
google-cloud>=0.34.0,<1.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==3.0.1
Pillow==9.3.0
